Move path utils to separate file
